MB-68266: Add explicit narrowing conversions in protocol code
Replace implicit narrowing conversions with explicit gsl::narrow calls
to satisfy -Wnarrowing compiler warnings:

 - client_connection.cc: Convert passphrase.size() (size_t) to int for
   OpenSSL PEM password callback return value

 - client_connection.cc: Convert timeout.count() and readTimeout.count()
   (chrono rep type) to uint32_t for folly::AsyncTimeout::scheduleTimeout

 - client_mcbp_commands.cc: Convert timeLimit.count() (int64_t) to
   uint32_t for RangeScanContinuePayload constructor

 - frameinfo.cc: Convert shift result (int-promoted from uint8_t << 4)
   back to uint8_t for frame encoding

All conversions are guaranteed safe by context: passphrase sizes are
limited, timeouts are reasonable values, and frame IDs are <= 0xf.

Reviewed-on: https://review.couchbase.org/c/kv_engine/+/244291
Tested-by: Build Bot <build@couchbase.com>

+++ protocol/connection/client_connection.cc
@@ -651,7 +651,7 @@ static int my_pem_password_cb(char* buf, int size, int, void* userdata) {
     }
 
     std::ranges::copy(passphrase, buf);
-    return passphrase.size();
+    return gsl::narrow<int>(passphrase.size());
 }
 
 void MemcachedConnection::connect() {
@@ -1175,10 +1175,10 @@ void MemcachedConnection::recvFrame(Frame& frame,
         auto timeoutvalue = readTimeout;
         if (readTimeout.count() == 0) {
             // none specified, use default
-            tmo.scheduleTimeout(timeout.count(), {});
+            tmo.scheduleTimeout(gsl::narrow<uint32_t>(timeout.count()), {});
             timeoutvalue = timeout;
         } else {
-            tmo.scheduleTimeout(readTimeout.count(), {});
+            tmo.scheduleTimeout(gsl::narrow<uint32_t>(readTimeout.count()), {});
         }
         asyncSocket->setReadCB(asyncReadCallback.get());
         if (!eventBase->loop()) {
